unscheduled-tasks: fix task count logic
the task count for unscheduled-task was shown wrong.

This is because the task count function was not taking care
if the task is completed or not.
Also, the plugin wasn't notifying for title changes when the task
count changed.

This commit fixes task count by counting only non completed unsheduled
tasks and notifying the application to change title on task count change.

https://bugzilla.gnome.org/show_bug.cgi?id=771906