task-list-view: only save task after setting the default date
When we add a new task from Today view, it doesn't appear immediately
because the due date of the task is being set ~after~ adding it to the
list (and not saving again to update it).

Fix that by setting the due date before saving the task to the tasklist.