Add a "todo" keyword to the desktop file.
"Todo" is a common keyword to describe a task management software, but
the software is called "To do" with the 2 original words split. Which is
fine except that it means that searching "todo" does not always return
"GNOME To Do". In particular, it won't work in the software center. It
does work in GNOME Shell, yet only as a side effect of the executable
being called `gnome-todo`. But then it will fail when packaged as
Flatpak which replaces the Exec value.