todo-txt: set default source when todo.txt source is not set
Earlier default source was set when the self->source was NULL.
but this aprroach is wrong as self->source is set to
g_settings_get_string (self->settings, "file") during initialization,
as g_settings_get_string return a empty string if the source is
never set.

So now set default source is called when the self->source is empty
and not NULL. Due to this the source is never set on first start
and hence the plugin does not loads.

https://bugzilla.gnome.org/show_bug.cgi?id=778583