todo.txt plugin: move to GFile API
Currently gtd-plugin-todo-txt has a source element in the struct
which stores the todo.txt path. This is not needed and the code
should make use of GFile API only.
so the source variable has been removed from the structure in both
provider and plugin and now we only use source_file.

https://bugzilla.gnome.org/show_bug.cgi?id=778583