unscheduled-panel: fix task loss issue
The tasklist used by the unscheduled panel does not have an associated provider.
This generates an error when attempting to create a task. In order to fix this,
do not explicitly set a task list for the view. In this way, the default
provider will be used, just like in the other plugins.

https://bugzilla.gnome.org/show_bug.cgi?id=773836