task-row: fix flickering when completing a task
When a task is completed, we change the GtdTask:complete property
immediately before starting the hide animation. This is an issue
since changing the :complete property causes the task list view to
be resorted, making the task go to the bottom of the list while
hiding.

Fix that by only setting the :complete property after the hide
animation is finished.