gtd-plugin-dialog: always open in list panel.
When we close the plugin dialog, instead of destroy it we hide
it, so when we open another time the plugin dialog, it shows the last
panel that was visible when we closed the dialog.

The correct behavoir should be that when we click show extensions in
the menu, always should appear the list of extensions.

https://bugzilla.gnome.org/show_bug.cgi?id=778588