ui: change default window size to show 3 columns
When restoring window from full screen for the first time, the windows
becomes a little too small for displaying three columns but remains too
big for two columns. Stays this way until the user resizes the window.

The default size was changed from 768 to 800 to accommodate the three
task list columns.

https://bugzilla.gnome.org/show_bug.cgi?id=777551