scheduled-panel: all labels show in scheduled view
The panel was being sorted, but seemed to place tasks in random groups

The problem is that both dates and times were set for each task. This
breaks the labeling, as single day gaps between tasks could be
interpreted as less than a day due to the fact that only whole days are
counted in the current mechanism. This groups certain tasks together,
which places some under the wrong header.

The fix is to seperate the working date difference system from the
mechanism which adds labels to the far right side of the task. Since
this works for labelling the task as "Tomorrow" or "Today", it is
suitable for also managing which headers should be set for each task.

https://bugzilla.gnome.org/show_bug.cgi?id=774034