todo-txt: create source file if it is deleted
This fixes the issue if the source_file that is being used is
deleted by the user when To Do is not running, in which case
earlier would have given an error as although the source contains
the uri to the file but the file is now present at the location.

The fix is to check if file exists and if not create one.

https://bugzilla.gnome.org/show_bug.cgi?id=779979