list-selector-grid-item don't update thumbnail on dispose
Rendering of list_selector_grid_item with NULL list reference causes
segmentation fault errors. While updating thumbnail of grid_item,
check is not imposed if list has been disposed.

The problem is that grid_item_update_thumbnail is being called even if
the list reference has been disposed.When provider is removed all the
list that it stores gets deleted.after disposal of item->list, grid item
widget flags changes which call the update_thumbnail which calls
render_thumbnail. This should only be called if list has not been
disposed as thumbnail will only be updated for non null list items in
the grid.

The fix is to check if the gtd-list-selector-item has a reference to valid
list and if true call the gtd_list_selector_grid_item__update_thumbnail
else do not call the update_thumbnail.

https://bugzilla.gnome.org/show_bug.cgi?id=778839