task-list-view: replaced "Done (0)" with "Done" in the done_label
In the list of tasks we have the label "Done (X)", after clicking
back_button we will get "Done (0)" for a few moments before the
page dissapears.

For fixing this bug we will use two variants of this label: the
first one for complete_tasks > 0, the second for complete_tasks = 0.

https://bugzilla.gnome.org/show_bug.cgi?id=781348