background: add a plugin to run To Do on background
As part of a fluid experience, it's nice to run GNOME
To Do in background for a few reasons.

First, because it's bad to keep reloading the tasklists
every time you reopen the app. This ultimately can cause
our Google keys to reach the daily limits.

Seconds, because now GNOME To Do can load when the session
starts up, and show a nice notification of how many tasks
do you have.

Third, because it's cool.