list-selector-list: Add ability to ctrl-click select
Currently entering selection mode is possible through clicking the 'selection
mode' button or selecting a list item (checkbox).

According to the style guide it's recomended to enable selection mode by
control/shift-clicking an item.

To remedy this we now allow ctrl-left-click on list items to enter
selection mode and mark them as selected.

https://bugzilla.gnome.org/show_bug.cgi?id=778675