eds: fix crash when changing default provider
When we change the default ESource in ESourceRegistry, it
emits a notify:: signal that we listen to. The problem is
that many GtdProviderEds instances were listening to the
same ESourceRegistry signal, without checking the if the
provider is itself.

This leads to a crash when selecting a tasklist from another
provider than the default. For example, by creating a new task
at an Online Account list, when the default is the local one.

Fix that by always checking if the provider of the new default
task list is itself, and adjusting the order of methods when
setting the default task list.