plugin_todo_txt: rework on plugin
Currently, the source is set after the plugin object is initialized,
but this should be done in init itself so the first the plugin is
laoded the provider is automatically loading and we dont have to send a
provider-added signal.
Other Changes are addition of function to set source, this helps in
code redundancy and make the code easier and better.

https://bugzilla.gnome.org/show_bug.cgi?id=781079